requirements-lock.txt
# NISystem Pinned Dependencies - Auto-generated from venv
# Regenerate with: pip freeze > requirements-lock.txt
# Last updated: 2026-02-01

altgraph==0.17.5
bcrypt==5.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deprecation==2.1.0
hightime==1.0.0
idna==3.11
iniconfig==2.3.0
nidaqmx==1.3.0
nitypes==1.1.0.dev1
numpy==2.4.0
packaging==25.0
paho-mqtt==2.1.0
pefile==2024.8.26
pluggy==1.6.0
psutil==7.2.1
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pymodbus==3.11.4
pyserial==3.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.10.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
